+++ src/components/AutosizeCodeEditor/Toolbar.tsx
@@ -1,0 +1,195 @@
+import { css, cx } from '@emotion/css';
+import { CodeEditorMonacoOptions, InlineField, InlineFieldRow, PageToolbar, ToolbarButton, useStyles2 } from '@grafana/ui';
+import type * as monacoType from 'monaco-editor/esm/vs/editor/editor.api';
+import React, { useEffect, useState } from 'react';
+
+import { TEST_IDS } from '../../constants';
+import { getStyles } from './AutosizeCodeEditor.styles';
+
+/**
+ * Properties
+ */
+interface Props {
+  /**
+   * Monaco editor instance
+   */
+  monacoEditor: monacoType.editor.IStandaloneCodeEditor | null;
+
+  /**
+   * Set modal open state
+   */
+  setIsOpen: (isOpen: boolean) => void;
+
+  /**
+   * Current editor value
+   */
+  editorValue: string;
+
+  /**
+   * Whether this toolbar is inside the modal
+   */
+  isModal?: boolean;
+
+  /**
+   * Set monaco options
+   */
+  setCurrentMonacoOptions: (options: CodeEditorMonacoOptions) => void;
+
+  /**
+   * Current monaco options
+   */
+  currentMonacoOptions: CodeEditorMonacoOptions | undefined;
+
+  /**
+   * Whether mini map is shown
+   */
+  isShowMiniMap: boolean | undefined;
+
+  /**
+   * Set mini map visibility
+   */
+  setIsShowMiniMap: (show: boolean | ((prev: boolean) => boolean)) => void;
+
+  /**
+   * Whether editor is read-only
+   */
+  readOnly?: boolean;
+}
+
+/**
+ * Toolbar
+ */
+export const Toolbar: React.FC<Props> = ({
+  monacoEditor,
+  setIsOpen,
+  editorValue,
+  isModal = false,
+  setCurrentMonacoOptions,
+  currentMonacoOptions,
+  isShowMiniMap,
+  setIsShowMiniMap,
+  readOnly,
+}) => {
+  /**
+   * Styles
+   */
+  const styles = useStyles2(getStyles);
+
+  /**
+   * Copy/Paste feedback message
+   */
+  const [feedbackMessage, setFeedbackMessage] = useState('');
+
+  /**
+   * Clear feedback message after 1 second
+   */
+  useEffect(() => {
+    if (feedbackMessage) {
+      const timer = setTimeout(() => {
+        setFeedbackMessage('');
+      }, 1000);
+      return () => clearTimeout(timer);
+    }
+    return;
+  }, [feedbackMessage]);
+
+  return (
+    <PageToolbar
+      buttonOverflowAlignment="right"
+      className={styles.toolbar}
+      forceShowLeftItems={true}
+      leftItems={[
+        <ToolbarButton
+          key="code-editor-button-modal"
+          tooltip={isModal ? 'Collapse code editor' : 'Expand code editor'}
+          icon={isModal ? 'compress-arrows' : 'expand-arrows-alt'}
+          iconSize="lg"
+          onClick={() => setIsOpen(!isModal)}
+          data-testid={TEST_IDS.codeEditor.modalButton(isModal ? 'modal-close' : 'modal-open')}
+        />,
+      ]}
+    >
+      <InlineFieldRow className={styles.copyPasteSection}>
+        <ToolbarButton
+          className={styles.copyPasteIcon}
+          tooltip="Copy code"
+          icon="file-blank"
+          iconSize="lg"
+          onClick={() => {
+            navigator.clipboard.writeText(editorValue).then(() => {
+              setFeedbackMessage('Copied!');
+            });
+          }}
+          data-testid={TEST_IDS.codeEditor.copyButton}
+        />
+        <ToolbarButton
+          disabled={readOnly}
+          className={styles.copyPasteIcon}
+          tooltip={readOnly ? 'Cannot edit in read-only mode' : 'Paste code'}
+          icon="file-alt"
+          iconSize="lg"
+          onClick={async () => {
+            if (monacoEditor) {
+              const selection = monacoEditor.getSelection();
+              const clipboardText = await navigator.clipboard.readText();
+              const range = {
+                startLineNumber: selection?.startLineNumber || 1,
+                startColumn: selection?.startColumn || 1,
+                endLineNumber: selection?.endLineNumber || 1,
+                endColumn: selection?.endColumn || 1,
+              };
+              monacoEditor.executeEdits('clipboard', [
+                {
+                  range,
+                  text: clipboardText,
+                  forceMoveMarkers: false,
+                },
+              ]);
+              monacoEditor.focus();
+              setFeedbackMessage('Pasted!');
+            }
+          }}
+          data-testid={TEST_IDS.codeEditor.pasteButton}
+        />
+        <InlineField
+          className={cx(
+            styles.copyPasteText,
+            feedbackMessage
+              ? css`
+                  width: 45px;
+                `
+              : css`
+                  width: 10px;
+                `
+          )}
+          data-testid={TEST_IDS.codeEditor.copyPasteText}
+        >
+          <div className={cx(styles.text, feedbackMessage ? styles.textVisible : '')}>{feedbackMessage}</div>
+        </InlineField>
+      </InlineFieldRow>
+      <ToolbarButton
+        tooltip="Wrap code on new lines"
+        icon="wrap-text"
+        iconSize="lg"
+        variant={currentMonacoOptions?.wordWrap === 'on' ? 'active' : 'default'}
+        onClick={() => {
+          const wordWrap = currentMonacoOptions?.wordWrap === 'on' ? 'off' : 'on';
+          setCurrentMonacoOptions({ ...currentMonacoOptions, wordWrap });
+        }}
+        data-testid={TEST_IDS.codeEditor.wrapButton}
+      />
+      {editorValue && editorValue.length > 100 && (
+        <ToolbarButton
+          tooltip={isShowMiniMap ? 'Hide mini map' : 'Show mini map'}
+          icon="gf-movepane-right"
+          iconSize="lg"
+          variant={isShowMiniMap ? 'active' : 'default'}
+          onClick={() => {
+            setIsShowMiniMap((prev: boolean) => !prev);
+          }}
+          data-testid={TEST_IDS.codeEditor.miniMapButton}
+        />
+      )}
+    </PageToolbar>
+  );
+};