+++ src/utils/code-parameters.ts
@@ -11,7 +11,7 @@ import {
 import { LocationService } from '@grafana/runtime';
 import { TimeZone } from '@grafana/schema';
 import { CodeEditorSuggestionItemKind } from '@grafana/ui';
-import { CodeParameterItem, CodeParametersBuilder } from '@volkovlabs/components';
+import { CodeParameterItem, CodeParametersBuilder } from './code-parameters-builder';
 import handlebars from 'handlebars';
 // eslint-disable-next-line @typescript-eslint/naming-convention
 import MarkdownIt from 'markdown-it';